requirements.txt
Django
djangorestframework
celery
redis
pandas
xlrd
psycopg2-binary

pytest
pytest-django
